dom/base/test/test_bug505783.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=505783
-->
<head>
  <title>Test for Bug 505783</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=505783">Mozilla Bug 505783</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 505783 **/

var a = document.createElementNS("http://www.w3.org/1999/xhtml", "a");
a.setAttributeNS("http://www.w3.org/XML/1998/namespace", "xml:base", "http://example.org/");
a.href = "page";
is(a.href, "http://example.org/page", "xml:base not used when not in doc");
document.getElementById("content").appendChild(a);
is(a.href, "http://example.org/page", "xml:base not used when in doc");

</script>
</pre>
</body>
</html>
